print-me-clone/src/lib/lambda/domain_layer/requirements.txt
python-ulid
boto3